pyproject.toml
[tool.poetry]
name = "alphabot"
version = "0.1.0"
description = "AlphaBot Multi-Agent Trading System"
authors = ["AlphaBot Team"]
readme = "README.md"
packages = [{include = "alphabot"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7"
crewai = "^0.28.0"
polars = "^0.20.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
scipy = "^1.10.0"
scikit-learn = "^1.3.0"
riskfolio-lib = "^5.0.0"
# streamlit = "^1.32.0"  # Temporairement désactivé pour tests
redis = "^5.0.0"
duckdb = "^0.9.0"
ib-insync = "^0.9.86"
yfinance = "^0.2.18"
alpha-vantage = "^2.3.1"
finnhub-python = "^2.4.18"
transformers = "^4.36.0"
torch = "^2.1.0"
pyyaml = "^6.0.1"
mlflow = "^2.9.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
ruff = "^0.1.0"
mypy = "^1.7.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
aiohttp = "^3.8.0"
vectorbt = "^0.25.0"
pydantic-settings = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
ruff = "^0.1.0"
mypy = "^1.7.0"
jupyter = "^1.0.0"
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=alphabot",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
    "-v"
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py39"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "crewai.*",
    "redis.*",
    "ib_insync.*",
    "finnhub.*",
    "alpha_vantage.*",
    "riskfolio.*"
]
ignore_missing_imports = true